Move startup logic to HomeActivity to hopefully save some resources
