+++ wear/src/main/AndroidManifest.xml
@@ -27,15 +27,13 @@
             android:name="com.google.android.wearable.standalone"
             android:value="true" />
 
-        <activity android:name=".launch.LaunchActivity"
+        <activity android:name=".home.HomeActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".home.HomeActivity" />
         <activity android:name=".onboarding.OnboardingActivity" />
         <activity android:name=".onboarding.integration.MobileAppIntegrationActivity" />
         <activity android:name=".onboarding.authentication.AuthenticationActivity" />
