+++ wear/src/main/java/io/homeassistant/companion/android/PresenterModule.kt
@@ -6,10 +6,6 @@ import dagger.Provides
 import io.homeassistant.companion.android.home.HomePresenter
 import io.homeassistant.companion.android.home.HomePresenterImpl
 import io.homeassistant.companion.android.home.HomeView
-import io.homeassistant.companion.android.launch.LaunchPresenter
-import io.homeassistant.companion.android.launch.LaunchPresenterImpl
-import io.homeassistant.companion.android.launch.LaunchView
-import io.homeassistant.companion.android.onboarding.OnboardingActivity
 import io.homeassistant.companion.android.onboarding.OnboardingPresenter
 import io.homeassistant.companion.android.onboarding.OnboardingPresenterImpl
 import io.homeassistant.companion.android.onboarding.OnboardingView
@@ -23,15 +19,10 @@ import io.homeassistant.companion.android.onboarding.integration.MobileAppIntegr
 @Module(includes = [PresenterModule.Declaration::class])
 class PresenterModule {
 
-    private lateinit var launchView: LaunchView
     private lateinit var onBoardingView: OnboardingView
     private lateinit var authenticationView: AuthenticationView
     private lateinit var mobileAppIntegrationView: MobileAppIntegrationView
     private lateinit var homeView: HomeView
-
-    constructor(launchView: LaunchView) {
-        this.launchView = launchView
-    }
 
     constructor(onBoardingView: OnboardingView) {
         this.onBoardingView = onBoardingView
@@ -50,9 +41,6 @@ class PresenterModule {
     }
 
     @Provides
-    fun provideLaunchView() = launchView
-
-    @Provides
     fun provideOnboardingView() = onBoardingView
 
     @Provides
@@ -66,9 +54,6 @@ class PresenterModule {
 
     @Module
     interface Declaration {
-
-        @Binds
-        fun bindLaunchPresenter(presenter: LaunchPresenterImpl): LaunchPresenter
 
         @Binds
         fun bindOnboardingPresenter(presenter: OnboardingPresenterImpl): OnboardingPresenter
