+++ wear/src/main/java/io/homeassistant/companion/android/home/HomeView.kt
@@ -4,5 +4,6 @@ interface HomeView {
     fun showHomeAssistantVersion(version: String)
     fun showEntitiesCount(count: Int)
 
-    fun displayLaunchView()
+    fun displayOnBoarding()
+    fun displayMobileAppIntegration()
 }
