--- wear/src/main/java/io/homeassistant/companion/android/launch/LaunchActivity.kt
@@ -1,62 +1,0 @@
-package io.homeassistant.companion.android.launch
-
-import android.content.Context
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import io.homeassistant.companion.android.DaggerPresenterComponent
-import io.homeassistant.companion.android.PresenterModule
-import io.homeassistant.companion.android.common.dagger.GraphComponentAccessor
-import io.homeassistant.companion.android.home.HomeActivity
-import io.homeassistant.companion.android.onboarding.OnboardingActivity
-import io.homeassistant.companion.android.onboarding.integration.MobileAppIntegrationActivity
-import javax.inject.Inject
-
-class LaunchActivity : AppCompatActivity(), LaunchView {
-
-    @Inject
-    lateinit var presenter: LaunchPresenter
-
-    companion object {
-        private const val TAG = "LaunchActivity"
-
-        fun newInstance(context: Context): Intent {
-            return Intent(context, LaunchActivity::class.java)
-        }
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        DaggerPresenterComponent
-            .builder()
-            .appComponent((application as GraphComponentAccessor).appComponent)
-            .presenterModule(PresenterModule(this))
-            .build()
-            .inject(this)
-
-        presenter.onViewReady()
-    }
-
-    override fun displayHome() {
-        startActivity(HomeActivity.newInstance(this))
-        finish()
-    }
-
-    override fun displayOnBoarding() {
-        val intent = OnboardingActivity.newInstance(this)
-        startActivity(intent)
-        finish()
-    }
-
-    override fun displayMobileAppIntegration() {
-        val intent = MobileAppIntegrationActivity.newInstance(this)
-        startActivity(intent)
-        finish()
-    }
-
-    override fun onDestroy() {
-        presenter.onFinish()
-        super.onDestroy()
-    }
-}
