--- wear/src/main/java/io/homeassistant/companion/android/launch/LaunchPresenter.kt
@@ -1,10 +1,0 @@
-package io.homeassistant.companion.android.launch
-
-interface LaunchPresenter {
-
-    fun onViewReady()
-
-    fun setSessionExpireMillis(value: Long)
-
-    fun onFinish()
-}
