--- wear/src/main/java/io/homeassistant/companion/android/launch/LaunchView.kt
@@ -1,9 +1,0 @@
-package io.homeassistant.companion.android.launch
-
-interface LaunchView {
-    fun displayHome()
-
-    fun displayOnBoarding()
-
-    fun displayMobileAppIntegration()
-}
