+++ wear/src/main/res/layout/activity_home.xml
@@ -1,62 +1,72 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/dark_grey"
-    android:fillViewport="true"
-    tools:context=".home.HomeActivity"
-    tools:deviceIds="wear">
-
-    <LinearLayout
+    android:layout_height="match_parent">
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_horizontal"
-        android:orientation="vertical"
-        android:padding="8dp">
+        android:layout_height="match_parent"
+        android:background="@color/dark_grey"
+        android:fillViewport="true"
+        tools:context=".home.HomeActivity"
+        tools:deviceIds="wear">
 
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="84dp"
-            android:layout_height="84dp"
-            android:contentDescription="@string/app_name"
-            android:src="@drawable/app_icon" />
-
-        <TextView
-            android:id="@+id/text"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:text="@string/app_name"
-            android:textColor="@android:color/white"
-            android:textSize="12sp" />
+            android:gravity="center_horizontal"
+            android:orientation="vertical"
+            android:padding="8dp">
 
-        <TextView
-            android:id="@+id/txt_version"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:textColor="@android:color/white"
-            android:textSize="12sp"
-            android:text="@string/loading"/>
+            <ImageView
+                android:id="@+id/imageView"
+                android:layout_width="84dp"
+                android:layout_height="84dp"
+                android:contentDescription="@string/app_name"
+                android:src="@drawable/app_icon" />
 
-        <TextView
-            android:id="@+id/txt_entities"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="5dp"
-            android:textColor="@android:color/white"
-            android:textSize="12sp"
-            android:text="@string/loading"/>
+            <TextView
+                android:id="@+id/text"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="5dp"
+                android:text="@string/app_name"
+                android:textColor="@android:color/white"
+                android:textSize="12sp" />
 
-        <com.google.android.material.button.MaterialButton
-            android:id="@+id/btn_logout"
-            android:layout_height="48dp"
-            android:layout_width="wrap_content"
-            android:layout_margin="8dp"
-            app:cornerRadius="24dp"
-            android:backgroundTint="@color/colorWarning"
-            android:text="@string/logout" />
-    </LinearLayout>
-</ScrollView>
+            <TextView
+                android:id="@+id/txt_version"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="5dp"
+                android:textColor="@android:color/white"
+                android:textSize="12sp"
+                android:text="@string/loading"/>
+
+            <TextView
+                android:id="@+id/txt_entities"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="5dp"
+                android:textColor="@android:color/white"
+                android:textSize="12sp"
+                android:text="@string/loading"/>
+
+            <com.google.android.material.button.MaterialButton
+                android:id="@+id/btn_logout"
+                android:layout_height="48dp"
+                android:layout_width="wrap_content"
+                android:layout_margin="8dp"
+                app:cornerRadius="24dp"
+                android:backgroundTint="@color/colorWarning"
+                android:text="@string/logout" />
+        </LinearLayout>
+    </ScrollView>
+    <io.homeassistant.companion.android.util.LoadingView
+        android:id="@+id/loading_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:loading_text="@string/checking_connection"
+        android:visibility="visible" />
+</FrameLayout>
